proyecto_pdf/requirements.txt
PyPDF2==3.0.1
openai==1.3.0
python-dotenv==1.0.0
reportlab==4.0.7
Pillow==10.1.0
